requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-filter==25.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.16.0
docopt==0.6.2
idna==3.10
requests==2.32.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20250402
typing_extensions==4.13.2
urllib3==2.4.0
yarg==0.1.10
